+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Par`'s `getPM()` function (also used to get the transverse velocity) parses `PMELONG` and `PMELAT`.
 - Added a `nonnegative=True` flag to `Archive`'s `getDynamicSpectrum()`. By default, if provided a template shape to match against the data, this will enforce non-negative (flux) dynamic spectrum values
 - `DynamicSpectrum`'s `imshow()` now allows a `cbarlabel` argument to provide a basic label to the colorbar, and a `badcolor` argument as an additional method to recolor instead of only `alpha`.
+- `Par` can now return an SWX model timeseries with `getSWX()`.
+
 
 ### Fixed